+++ docs/CHANGELOG.md
@@ -2,16 +2,26 @@
     CHANGELOG
 </h1>
 
-**MangDL v.0.0.0-alpha.0**
+## **MangDL v.0.0.0-alpha.1**
+
+**ADDED**
+
+-  <a target="_blank" href="https://danke.moe">Danke.moe</a> provider
+
+## **MangDL v.0.0.0-alpha.0**
 
 START OF SEMANTIC VERSIONING.
 
-**ADDED**
-- Hachirumi provider
+### **ADDED**
 
-**CHANGED**
+<a target="_blank" href="https://hachirumi.com">Hachirumi.com</a> provider
+
+### **CHANGED**
+
 - An intermediary script between the provider scripts and the cli has been made, which can also be used thru programmatic usage.
+
 Ex.:
+
 ```python
 from mangdl import Provider
 from mangdl.API.Base import Search
@@ -21,7 +31,8 @@ for i in md.search(Search("sachi-iro no one room")):
     print(i)
 ```
 
-**REMOVED**
+### **REMOVED**
+
 - Saving and loading configs
 
 <sub>